Allow request timeout modification in server
With this change we want to make the request timeout of the server configurable. This is needed because server request timeout has changed from Node.js v18. Now it is set to 300000 milliseconds (5 minutes) by default. See docs: https://fek.io/blog/what-is-new-in-node-js-18/.

Our current use case is that we have an upload asset endpoint that takes some time to upload a file, so with this default timeout the request was failing with a "Request timeout" error. That's why we need to customise this timeout value.